+++ data/crac/crac-io/crac-io-json/src/main/java/com/powsybl/openrao/data/crac/io/json/deserializers/InjectionRangeActionArrayDeserializer.java
@@ -10,7 +10,7 @@ package com.powsybl.openrao.data.crac.io.json.deserializers;
 import com.powsybl.iidm.network.Network;
 import com.powsybl.openrao.commons.OpenRaoException;
 import com.powsybl.openrao.data.crac.io.json.JsonSerializationConstants;
-import com.powsybl.openrao.data.crac.io.commons.ucte.InjectionRangeActionHelper;
+import com.powsybl.openrao.data.crac.io.commons.iidm.IidmInjectionHelper;
 import com.powsybl.openrao.data.crac.api.Crac;
 import com.powsybl.openrao.data.crac.api.rangeaction.InjectionRangeActionAdder;
 import com.fasterxml.jackson.core.JsonParser;
@@ -47,7 +47,7 @@ public final class InjectionRangeActionArrayDeserializer {
                 }
             }
 
-            double initialSetpoint = InjectionRangeActionHelper.getCurrentSetpoint(network, injectionDistributionKeys);
+            double initialSetpoint = IidmInjectionHelper.getCurrentSetpoint(network, injectionDistributionKeys);
             injectionRangeActionAdder.withInitialSetpoint(initialSetpoint);
             injectionRangeActionAdder.add();
